add fqdn support
