+++ src/azure-cli/azure/cli/command_modules/acr/_params.py
@@ -135,6 +135,38 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('read_enabled', help='Indicates whether read operation is allowed.', arg_type=get_three_state_flag())
         c.argument('write_enabled', help='Indicates whether write or delete operation is allowed.', arg_type=get_three_state_flag())
 
+    with self.argument_context('acr manifest') as c:
+        c.argument('registry_name', options_list=['--registry', '-r'], help='The name of the container registry. You can configure the default registry name using `az configure --defaults acr=<registry name>`', completer=get_resource_name_completion_list(REGISTRY_RESOURCE_TYPE), configured_default='acr', validator=validate_registry_name)
+        c.argument('top', type=int, help='Limit the number of items in the results.')
+        c.argument('orderby', help='Order the items in the results. Default to alphabetical order of names.', arg_type=get_enum_type(['time_asc', 'time_desc']))
+        c.argument('delete_enabled', help='Indicates whether delete operation is allowed.', arg_type=get_three_state_flag())
+        c.argument('list_enabled', help='Indicates whether this item shows in list operation results.', arg_type=get_three_state_flag())
+        c.argument('read_enabled', help='Indicates whether read operation is allowed.', arg_type=get_three_state_flag())
+        c.argument('write_enabled', help='Indicates whether write or delete operation is allowed.', arg_type=get_three_state_flag())
+        c.argument('repository', options_list=['--name', '-n'])
+        c.argument('manifest_id', options_list=['--name', '-n'])
+
+    with self.argument_context('acr manifest show') as c:
+        c.positional('id', nargs='*', default=None)
+
+    with self.argument_context('acr manifest list') as c:
+        c.positional('id', nargs='*', default=None)
+
+    with self.argument_context('acr manifest delete') as c:
+        c.positional('id', nargs='*', default=None)
+
+    with self.argument_context('acr manifest show-metadata') as c:
+        c.positional('id', nargs='*', default=None)
+
+    with self.argument_context('acr manifest list-metadata') as c:
+        c.positional('id', nargs='*', default=None)
+
+    with self.argument_context('acr manifest update-metadata') as c:
+        c.positional('id', nargs='*', default=None)
+
+    with self.argument_context('acr manifest list-referrers') as c:
+        c.positional('id', nargs='*', default=None)
+
     with self.argument_context('acr repository untag') as c:
         c.argument('image', options_list=['--image', '-t'], help="The name of the image. May include a tag in the format 'name:tag'.")
