+++ src/azure-cli/azure/cli/command_modules/acr/commands.py
@@ -82,6 +82,10 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         operations_tmpl='azure.cli.command_modules.acr.repository#{}'
     )
 
+    acr_manifest_util = CliCommandType(
+        operations_tmpl='azure.cli.command_modules.acr.manifest#{}'
+    )
+
     acr_webhook_util = CliCommandType(
         operations_tmpl='azure.cli.command_modules.acr.webhook#{}',
         table_transformer=webhook_output_format,
@@ -207,7 +211,7 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('delete', 'acr_repository_delete')
         g.command('untag', 'acr_repository_untag')
 
-    with self.command_group('acr repository manifest', acr_repo_util, is_preview=True) as g:
+    with self.command_group('acr manifest', acr_manifest_util, is_preview=True) as g:
         g.show_command('show', 'acr_repository_show_manifest', table_transformer=show_manifest_output_format)
         g.command('list', 'acr_repository_list_manifests', table_transformer=list_manifest_output_format)
         g.command('delete', 'acr_repository_delete_manifests')
@@ -215,8 +219,6 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('list-metadata', 'acr_repository_list_manifest_metadata')
         g.command('update-metadata', 'acr_repository_update_manifest_metadata')
         g.command('list-referrers', 'acr_repository_list_manifest_referrers', table_transformer=list_references_output_format)
-
-
 
     with self.command_group('acr webhook', acr_webhook_util) as g:
         g.command('list', 'acr_webhook_list')
